+++ api.go
@@ -12,6 +12,8 @@ import (
 	"strconv"
 	"strings"
 	"flag"
+	"sync"
+	"sort"
 )
 
 /* worker related constants */
@@ -22,8 +24,9 @@ const (
 
 /* logging related constants */
 const (
-	APP_DIR = "polyglot-task-by-JK-aka-mrl5"
+	APP_DIR = "polyglot-task-by-JK-aka-mrl5-v20"
 	INPUT_LOG_FILE = "input.log"
+	REQUESTS_LOG_FILE = "requests.log"
 	APPDIR_PERMISSION = 0750   //owner can do anything with directory, group member can list directory and see rights, others wont see dir content
 	LOGFILES_PERMISSION = 0640 //RW for owner, R for group, nothing for others (0640)
 )
@@ -31,35 +34,21 @@ const (
 var endpointUUID string
 var logsDir = ""
 var pathToInputLog string
+var pathToRequestsLog string
 
 func setFlags() {
 	/* flags related constants */
 	const (
-		DEFAULT_UUID = "internal"
-		UUID_USAGE = "Universally Unique IDentifier of an endpoint request"
+		defaultUUID = "internal"
+		uuidUsage = "Universally Unique IDentifier of the endpoint request"
 	)
 
 	/* UUID */
-	flag.StringVar(&endpointUUID, "uuid", DEFAULT_UUID, UUID_USAGE)
-	flag.StringVar(&endpointUUID, "u", DEFAULT_UUID, UUID_USAGE+" (shorthand)")
+	flag.StringVar(&endpointUUID, "uuid", defaultUUID, uuidUsage)
+	flag.StringVar(&endpointUUID, "u", defaultUUID, uuidUsage+" (shorthand)")
 
 	/* parse all defined flags */
 	flag.Parse()
-}
-
-func checkInput() (int, error) {
-	fmt.Println(os.Args)
-	fmt.Println(len(flag.Args()))
-	fmt.Println(flag.Args())
-	fmt.Println(endpointUUID)
-	switch len(flag.Args()) {
-	case 1:
-		return 0, nil
-	case 0:
-		return -1, errors.New("API error: expected an argument")
-	default:
-		return -2, errors.New("API error: only one argument is accepted")
-	}
 }
 
 func checkError(e error) {
@@ -73,57 +62,130 @@ func checkEnvironment(usrDir string) {
 		logsDir = path.Join(usrDir, APP_DIR)
 	}
 	pathToInputLog = path.Join(logsDir, INPUT_LOG_FILE)
+	pathToRequestsLog = path.Join(logsDir, REQUESTS_LOG_FILE)
 
-	/* create logs dir if doesn't exist, ignore error when exists */
+	/* create logs dir if doesn't exist, ignore error when it exists */
 	_ = os.Mkdir(logsDir, APPDIR_PERMISSION)
 
-	/* check if log files exists */
+	/* check if log files exist, create when they don't */
 	os.OpenFile(pathToInputLog, os.O_RDONLY|os.O_CREATE, LOGFILES_PERMISSION)
+	os.OpenFile(pathToRequestsLog, os.O_RDONLY|os.O_CREATE, LOGFILES_PERMISSION)
 }
 
-func logInput(input string) {
+func checkInput(noOfArgs int) (int, error) {
+	switch noOfArgs {
+	default:
+		return 0, nil
+	case 0:
+		return -1, errors.New("API error: expected at least one argument")
+	}
+}
+
+func logInput(input string, elapsedProcessTime string) {
 	f, err := os.OpenFile(pathToInputLog, os.O_APPEND|os.O_WRONLY, LOGFILES_PERMISSION)
 	checkError(err)
 	defer f.Close()
 
-	logLine := endpointUUID + "\t" + input
+	/* uuid elapsed-time input */
+	logLine := endpointUUID + "\t" + elapsedProcessTime + "\t" + input + "\n"
 	if _, err = f.WriteString(logLine); err != nil {
 		panic(err)
 	}
 }
 
-func executeWorker(pathToWorker string, argument string) (error, []byte) {
+func logRequest(requestStart time.Time, noOfArguments int) {
+	f, err := os.OpenFile(pathToRequestsLog, os.O_APPEND|os.O_WRONLY, LOGFILES_PERMISSION)
+	checkError(err)
+	defer f.Close()
+
+	/* date uuid elapsed-time number-of-args */
+	logLine := "[" + requestStart.UTC().String() + "]\t" + endpointUUID + "\t"  +
+		strconv.FormatFloat(time.Since(requestStart).Seconds(), 'f', 3, 64) + "\t" +
+		strconv.Itoa(noOfArguments) + "\n"
+	if _, err = f.WriteString(logLine); err != nil {
+		panic(err)
+	}
+}
+
+
+/* goroutines framework */
+type workerReturn struct {
+	id int
+	err error
+	output []byte
+	elapsedTime time.Duration
+}
+
+type workerReturnArr []workerReturn
+var workerReturns workerReturnArr
+
+func (wr workerReturnArr) Len() int {
+	return len(wr)
+}
+
+func (wr workerReturnArr) Less(i, j int) bool {
+	return wr[i].id < wr[j].id
+}
+
+func (wr workerReturnArr) Swap(i, j int) {
+	wr[i], wr[j] = wr[j], wr[i]
+}
+/* /goroutines framework */
+
+func executeWorker(pathToWorker string, argument string, id int, wg *sync.WaitGroup) {
+	start := time.Now()
 	workerProcess := exec.Command(pathToWorker, argument)
 	output, err := workerProcess.Output()
-	return err, output
+	/* add return to the array of returns */
+	workerReturns = append(workerReturns, workerReturn{id, err, output, time.Since(start)})
+	wg.Done()
 }
 
 func main() {
 	start := time.Now()
-	var elapsedTime string
 	setFlags()
 
 	/* path to the directory of binary version of this program */
 	workDir, workDirErr := filepath.Abs(filepath.Dir(os.Args[0]))
 	checkError(workDirErr)
+	pathToWorker := path.Join(workDir, WORKER)
 
 	usrEnv, usrEnvErr := user.Current()
 	checkError(usrEnvErr)
 	checkEnvironment(usrEnv.HomeDir)
 
-	/* check input syntax */
-	if _, argErr := checkInput(); argErr == nil {
-		pathToWorker := path.Join(workDir, WORKER)
-		err, result := executeWorker(pathToWorker, flag.Arg(0))
-		execTime := time.Since(start).Seconds()
-		elapsedTime = strconv.FormatFloat(execTime, 'f', 3, 64)
-		if err != nil {
-			fmt.Println(ERROR_MSG)
-		} else {
-			fmt.Println(strings.Trim(string(result), "\n") + ", " + elapsedTime)
+	noOfArgs := len(flag.Args())
+
+	if _, argErr := checkInput(noOfArgs); argErr == nil {
+		/* sync.WaitGroup waits for a group of goroutines to finish */
+		var wg sync.WaitGroup
+		wg.Add(noOfArgs)
+
+		for i := 0; i < noOfArgs; i++ {
+			go executeWorker(pathToWorker, flag.Arg(i), i, &wg)
 		}
+		/* wait until all executeWorker processes will end */
+		wg.Wait()
+		/* FIFO sorting (by id) */
+		sort.Sort(workerReturns)
+
+		for i := 0; i < noOfArgs; i++ {
+			var result string
+			formattedTime := strconv.FormatFloat(
+				workerReturns[i].elapsedTime.Seconds(), 'f', 3, 64)
+
+			if workerReturns[i].err != nil {
+				result = ERROR_MSG
+			} else {
+				result = strings.Trim(
+					string(workerReturns[i].output), "\n")
+			}
+			fmt.Println(result + ", " + formattedTime)
+			logInput(flag.Arg(i), formattedTime)
+		}
+
 	} else {
 		fmt.Println(argErr)
 	}
-	logInput(elapsedTime + "\t" + flag.Arg(0) + "\n")
+	logRequest(start, len(flag.Args()))
 }