+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 All **notable** changes to this project will be documented in this file.
 
 ## [v2.0-SNAPSHOT]
+- `endpoint.rb` calls API **once**
+- `api.go` uses **goroutines** to call `worker.py`
 - new `input.log` format: `<request UUID>	<execution time>	<input expression>`
+- new logfile: `requests.log`
 
 ##  v1.0 - 2018-10-02
 - first official release